+++ src/WebJobs.Extensions.OpenAI.CosmosDBSearch/CosmosDBSearchConfigOptions.cs
@@ -15,7 +15,7 @@ public class CosmosDBSearchConfigOptions
     public string IndexName { get; set; } = "openai-functions-index";
 
     // Application name for the client for tracking and logging
-    public string ApplicationName { get; set; } = "openai-function-appName";
+    public string ApplicationName { get; } = "OPENAI_FUNCTIONAPP";
 
     // Name of the field property which will contain the embeddings
     public string EmbeddingKey { get; set; } = "embedding";
@@ -34,6 +34,7 @@ public class CosmosDBSearchConfigOptions
     ///     Possible options are:
     ///         - vector-ivf
     ///         - vector-hnsw
+    ///         - vector-diskann
     /// </summary>
     public string Kind { get; set; } = "vector-ivf";
 
@@ -64,4 +65,22 @@ public class CosmosDBSearchConfigOptions
     /// the cost of speed.
     /// </summary>
     public int EfSearch { get; set; } = 40;
+
+    /// <summary>
+    /// Max number of neighbors. Only vector-diskann search supports this for now.
+    /// Default value is 32, range from 20 to 2048.
+    /// </summary>
+    public int MaxDegree { get; set; } = 32;
+
+    /// <summary>
+    /// l value for index building. Only vector-diskann search supports this for now.
+    /// Default value is 50, range from 10 to 500.
+    /// </summary>
+    public int LBuild { get; set; } = 50;
+
+    /// <summary>
+    /// l value for index searching. Only vector-diskann search supports this for now.
+    /// Default value is 40, range from 10 to 10000.
+    /// </summary>
+    public int LSearch { get; set; } = 40;
 }